backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "python-jose[cryptography]>=3.3.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "passlib>=1.7.4",
    "argon2-cffi>=21.3.0",
    "bcrypt>=4.0.0",
    "openai-agents>=0.6.5",
    "mcp-server>=0.1.4",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
    "pytest-asyncio>=1.3.0",
]
